refactored - BaseModule
