+++ configs/model/criterion/ce_loss.yaml
@@ -1,0 +1,1 @@
+_target_: torch.nn.CrossEntropyLoss
